% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeImbalance.R
\name{analyzeImbalance}
\alias{analyzeImbalance}
\title{Analyze Imbalance}
\usage{
analyzeImbalance(x, y, phi = NULL, thresh_rel = 0.5, rel_method = "PCHIP", ...)
}
\arguments{
\item{x}{feature matrix or dataframe. Only numeric variables for now.}

\item{y}{target variable.}

\item{phi}{Relevance values for each sample. Default is NULL. If NULL,
inverse of kernel density estimation is used.}

\item{thresh_rel}{threshold to determine rare values. Default is 0.5.}

\item{rel_method}{method for relevance function. Default is "PCHIP". Choices
are "PCHIP" and "density". Ignored if phi is given.}

\item{...}{relevance function settings.}
}
\value{
an list object which includes:
 \item{x_original}{original feature matrix}
 \item{y_original}{original target variable}
 \item{groups_original}{categories of original data}
 \item{phi}{relevance function values for y}
 \item{rel_model}{Details about relevance function. Can be used to calculate
 new relevance for test data.}
}
\description{
Gets information about imbalance ratios based on relevance function.
}
\details{
\code{analyzeImbalance} determines rare instances based on specified relevance
function. Then creates a group variable to indicate the characteristics of
respected instance. Function reorders datasets so output "groups" variable is
based on output "y_original". Imbalance ratios are calculated based on number of
samples in both upper and lower rare area. It is useful to see the
consequences of specified relevance function.
}
\examples{
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

analyzeImbalance(x = x, y = y)


}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
