% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSMOTER.R
\name{GSMOTER}
\alias{GSMOTER}
\title{Geometric SMOTE for Regression}
\usage{
GSMOTER(
  x,
  y,
  thresh_rel = 0.5,
  phi = NULL,
  perc_ov_lower = NULL,
  perc_ov_upper = NULL,
  perc_un = NULL,
  k = 5,
  alpha_sel = "combined",
  alpha_trunc = 0.5,
  alpha_def = 0.5,
  rel_method = "PCHIP",
  ...
)
}
\arguments{
\item{x}{feature matrix or dataframe. Only numeric variables for now.}

\item{y}{target variable.}

\item{thresh_rel}{threshold to determine rare values. Default is 0.5.}

\item{phi}{Relevance values for each sample. Default is NULL. If NULL,
inverse of kernel density estimation is used.}

\item{perc_ov_lower}{percentage of oversampling of lower rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_ov_upper}{percentage of oversampling of upper rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_un}{percentage of undersampling of non-rare samples. If NULL, it
is automatically determined by algorithm. Cannot be higher than 1.}

\item{k}{number of neighbors for links.}

\item{alpha_sel}{selection method. Can be "minority", "majority" or "combined".
Default is "combined".}

\item{alpha_trunc}{truncation factor. A numeric value in \eqn{[-1,1]}.
Default is 0.5.}

\item{alpha_def}{deformation factor. A numeric value in \eqn{[0,1]}.
Default is 0.5}

\item{rel_method}{method for relevance function. Default is "PCHIP". Choices
are "PCHIP" and "density". Ignored if phi is given.}

\item{...}{relevance function settings.}
}
\value{
an list object which includes:
 \item{x_new}{Balanced feature matrix}
 \item{y_new}{Balanced target variable}
 \item{groups_new}{categories of new data}
 \item{x_original}{original feature matrix}
 \item{y_original}{original target variable}
 \item{groups_original}{categories of original data}
 \item{x_syn}{Synthetic feature matrix}
 \item{y_syn}{Synthetic target variable}
 \item{phi}{relevance function values for y}
 \item{rel_model}{Details about relevance function. Can be used to calculate
 new relevance for test data.}
}
\description{
\code{GSMOTER} applies undersampling and oversampling to
imbalanced regression data sets.
}
\details{
Despite its name, it can both undersample and oversample imbalanced data sets.
It is used to better estimate the rare values in regression. Algorithm is not
exactly from the original paper, Camacho et al. (2022). Paper does not give
an algorithm and references to Douzas & Bacao (2019). Details are in
both papers.

There are three classes: lower rare, not rare and upper
rare. Lower rare class is samples that satisfy \eqn{\phi > treshold} and
\eqn{y < \tilde{y}}. Upper rare class is samples that satisfy
\eqn{\phi > treshold} and \eqn{y > \tilde{y}}. Other samples are considered
as not rare. \eqn{\tilde{y}} is median of \eqn{y}.
}
\examples{
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

m_GSMOTER <- GSMOTER(x = x, y = y)

plot(x, y)
plot(m_GSMOTER$x_new, m_GSMOTER$y_new)

}
\references{
Camacho, L., Douzas, G., & Bacao, F. (2022). Geometric SMOTE for regression.
Expert Systems with Applications, 193, 116387.

Douzas, G., & Bacao, F. (2019). Geometric SMOTE a geometrically enhanced
drop-in replacement for SMOTE. Information sciences, 501, 118-135.
}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
