% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_density.R
\name{relevance_density}
\alias{relevance_density}
\title{Relevance function based on density}
\usage{
relevance_density(y, y_new = y, h = NULL, method = "ucv")
}
\arguments{
\item{y}{target values.}

\item{y_new}{new target values for calculating relevance values. Default is y.}

\item{h}{bandwith value for kernel density estimation. Default is NULL. If
NULL, it is estimated based on \code{method}.}

\item{method}{method to be used in estimating bandwidth. Default is unbiased
cross validation.}
}
\value{
a vector of relevance values.
}
\description{
Relevance value generation using kernel density estimation
}
\details{
\code{relevance_density} is a relevance function. It is used to determine
rare and not rare values in a regression problem. It calculates kernel densities
and inverse of densities are considered relevance values.
}
\examples{

### regression dataset
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

### relevance values
rel <- relevance_density(y)

}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
