% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_PCHIP.R
\name{relevance_PCHIP}
\alias{relevance_PCHIP}
\title{Piecewise Cubic Hermite Interpolating Polynomials}
\usage{
relevance_PCHIP(
  y,
  y_new = y,
  coef = 1.5,
  S_points = NULL,
  S_phi = NULL,
  S_der = NULL
)
}
\arguments{
\item{y}{target values.}

\item{y_new}{new target values for calculating relevance values. Default is y.}

\item{coef}{coefficient for IQR outlier range. Default is 1.5.}

\item{S_points}{a numeric vector of length 3. They are control points in y.
Lower than first value are lower rare, higher than third value are upper rare
area. Middle value is median and have zero relevance. Default is NULL and
automatically decided based on IQR statistics.}

\item{S_phi}{relevance values for respected points given in S_points. Default
is NULL and automatically decided.}

\item{S_der}{derivative of relevance values for respected points given in
S_points. If extreme points are given in S_points, all three values should
be zero. Default is NULL, which means all values are zero.}
}
\value{
a vector of relevance values.
}
\description{
Relevance function using piecewise cubic Hermite interpolating
polynomials (PCHIP)
}
\details{
\code{relevance_PCHIP} is the most popular relevance function. It is used to
determine rare and not rare values in a regression problem. It calculates
interpolation values to determine weights of samples based on given bumps
(S_points).
}
\examples{

### regression dataset
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

### relevance values
rel <- relevance_PCHIP(y)

}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
