% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_sigmoid.R
\name{relevance_sigmoid}
\alias{relevance_sigmoid}
\title{Relevance function based on sigmoid}
\usage{
relevance_sigmoid(
  y,
  y_new = y,
  coef = 1.5,
  adj_points = NULL,
  delta = 1e-04,
  k = 0.5
)
}
\arguments{
\item{y}{target values.}

\item{y_new}{new target values for calculating relevance values. Default is y.}

\item{coef}{coefficient for IQR outlier range. Default is \eqn{1.5}.}

\item{adj_points}{a numeric vector of length 2. They are control points which
represents the points where relevance equals \eqn{0.5}. First value is the lower,
second is the upper \eqn{0.5}. Default is NULL and automatically decided based on
IQR statistics.}

\item{delta}{precision value. Default is \eqn{1e-4}.}

\item{k}{decay factor. Default is \eqn{0.1}.}
}
\value{
a vector of relevance values.
}
\description{
Relevance value generation using sigmoid function
}
\details{
\code{relevance_sigmoid} is a relevance function. It is used to determine
rare and not rare values in a regression problem. It calculates relevance
based on a modified sigmoid function (Torgo & Ribeiro, 2021).
}
\examples{
### regression dataset
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

### relevance values
rel <- relevance_sigmoid(y)

}
\references{
Torgo, L., & Ribeiro, R. (2009). Precision and recall for regression. In
Discovery Science: 12th International Conference, DS 2009, Porto, Portugal,
October 3-5, 2009 12 (pp. 332-346). Springer Berlin Heidelberg.
}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
