% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUS.R
\name{ROUS}
\alias{ROUS}
\title{Random over-undersampling}
\usage{
ROUS(
  x,
  y,
  phi = NULL,
  thresh_rel = 0.5,
  perc_ov_lower = NULL,
  perc_ov_upper = NULL,
  perc_un = NULL,
  rel_method = "PCHIP",
  ...
)
}
\arguments{
\item{x}{feature matrix or dataframe. Only numeric variables for now.}

\item{y}{target variable.}

\item{phi}{Relevance values for each sample. Default is NULL. If NULL,
inverse of kernel density estimation is used.}

\item{thresh_rel}{threshold to determine rare values. Default is 0.5.}

\item{perc_ov_lower}{percentage of oversampling of lower rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_ov_upper}{percentage of oversampling of upper rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_un}{percentage of undersampling of non-rare samples. If NULL, it
is automatically determined by algorithm. Cannot be higher than 1.}

\item{rel_method}{method for relevance function. Default is "PCHIP". Choices
are "PCHIP" and "density". Ignored if phi is given.}

\item{...}{relevance function settings.}
}
\value{
an list object which includes:
 \item{x_new}{Balanced feature matrix}
 \item{y_new}{Balanced target variable}
 \item{groups_new}{categories of new data}
 \item{x_original}{original feature matrix}
 \item{y_original}{original target variable}
 \item{groups_original}{categories of original data}
 \item{x_syn}{Synthetic feature matrix}
 \item{y_syn}{Synthetic target variable}
 \item{phi}{relevance function values for y}
 \item{rel_model}{Details about relevance function. Can be used to calculate
 new relevance for test data.}
}
\description{
\code{ROUS} applies random undersampling and oversampling to
imbalanced regression data sets.
}
\details{
Random over-undersampling (ROUS) can be used as random oversampling (ROS) and
random undersampling. Random oversampling copies randomly selected rare
instances and addes to data. Random udersampling randomly selects non-rare
instances and removes them.
}
\examples{
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

m_ROUS <- ROUS(x = x, y = y)

plot(x, y)
plot(m_ROUS$x_new, m_ROUS$y_new)

}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
