% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOGN.R
\name{SMOGN}
\alias{SMOGN}
\title{SMOGN}
\usage{
SMOGN(
  x,
  y,
  phi = NULL,
  thresh_rel = 0.5,
  k = 5,
  perc_ov_lower = NULL,
  perc_ov_upper = NULL,
  perc_un = NULL,
  rel_method = "PCHIP",
  pert = 0.02,
  ...
)
}
\arguments{
\item{x}{feature matrix or dataframe. Only numeric variables for now.}

\item{y}{target variable.}

\item{phi}{Relevance values for each sample. Default is NULL. If NULL,
inverse of kernel density estimation is used.}

\item{thresh_rel}{threshold to determine rare values. Default is 0.5.}

\item{k}{number of neighbors for links.}

\item{perc_ov_lower}{percentage of oversampling of lower rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_ov_upper}{percentage of oversampling of upper rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_un}{percentage of undersampling of non-rare samples. If NULL, it
is automatically determined by algorithm. Cannot be higher than 1.}

\item{rel_method}{method for relevance function. Default is "PCHIP". Choices
are "PCHIP" and "density".}

\item{pert}{effects the variance of noises. Default is 0.02, same as the paper
(Branco et al., 2017).}

\item{...}{relevance function settings.}
}
\value{
an list object which includes:
 \item{x_new}{Balanced feature matrix}
 \item{y_new}{Balanced target variable}
 \item{groups_new}{categories of new data}
 \item{x_original}{original feature matrix}
 \item{y_original}{original target variable}
 \item{groups_original}{categories of original data}
 \item{x_syn}{Synthetic feature matrix}
 \item{y_syn}{Synthetic target variable}
 \item{phi}{relevance function values for y}
 \item{rel_model}{Details about relevance function. Can be used to calculate
 new relevance for test data.}
}
\description{
\code{SMOTER} applies undersampling and oversampling to
imbalanced regression data sets.
}
\details{
SMOGN (Branco et al., 2017) generates synthetic data sometimes SMOTE-like
and sometimes by adding Gaussian noise. It compares the length of link to the
median distance to other rare samples. If the first is higher, SMOGN generates
by adding noise. If the second is higher, it generates sample on the link.
}
\examples{
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

m_SMOGN <- SMOGN(x = x, y = y)

plot(x, y)
plot(m_SMOGN$x_new, m_SMOGN$y_new)

}
\references{
Branco, P., Torgo, L., & Ribeiro, R. P. (2017, October). SMOGN: a
pre-processing approach for imbalanced regression. In First international
workshop on learning with imbalanced domains: Theory and applications
(pp. 36-50). PMLR.
}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
