% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSSE.R
\name{WSSE}
\alias{WSSE}
\title{Weighted Sum of Squared Error}
\usage{
WSSE(truth, pred, weight = NULL)
}
\arguments{
\item{truth}{true values.}

\item{pred}{predictions.}

\item{weight}{a numerical vector of length same as truth and pred. Default is
NULL. If NULL, it is equal to all samples are equally weighted.}
}
\value{
a numerical value.
}
\description{
Calculation of weighted sum of squared error for weighted
regression models
}
\details{
Calculates weighted sum of squared error. It can be used as a performance or
model selection metric in imbalanced regression.
}
\examples{

x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

x_train = x[1:70]
y_train = y[1:70]

x_test = x[71:100]
y_test = y[71:100]

### resampling with SMOTER to predict rare values better
m_SMOTER <- SMOTER(x = x_train, y = y_train)

data_train <- data.frame(x = m_SMOTER$x_new, y = m_SMOTER$y_new)
m_lm <- lm(y~., data = data_train)

pred_test <- predict(m_lm, newdata = data.frame(x = x_test))

### classical sum of squares
SSE_notweighted <- WSSE(truth = y_test, pred = pred_test)
SSE_notweighted

### weights for test data
m_rel <- relevance_PCHIP(y = y_train, y_new = y_test)

### weighted sum of squares
SSE_weighted <- WSSE(truth = y_test, pred = pred_test, weight = m_rel$rel)
SSE_weighted

}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
