% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedMetrics.R
\name{WSSE}
\alias{WSSE}
\alias{WMSE}
\alias{WMAD}
\alias{WMAPE}
\alias{MU}
\alias{NMU}
\alias{precisionReg}
\alias{recallReg}
\alias{Fbeta}
\title{Weighted Performance Metrics}
\usage{
WSSE(truth, pred, phi = NULL, ...)

WMSE(truth, pred, phi = NULL, ...)

WMAD(truth, pred, phi = NULL, ...)

WMAPE(truth, pred, phi = NULL, ...)

MU(truth, pred, phi = NULL, y_train, ...)

NMU(truth, pred, phi = NULL, y_train, ...)

precisionReg(truth, pred, phi = NULL, thresh = 0.5, y_train, p = 0.5, ...)

recallReg(truth, pred, phi = NULL, thresh = 0.5, y_train, p = 0.5, ...)

Fbeta(truth, pred, phi = NULL, thresh = 0.5, y_train, p = 0.5, beta = 1, ...)
}
\arguments{
\item{truth}{true values.}

\item{pred}{predictions.}

\item{phi}{a numerical vector of length same as truth and pred. Default is
NULL. If NULL, it is equal to all samples are equally weighted.}

\item{...}{not used for now.}

\item{y_train}{train target variable.}

\item{thresh}{threshold for rare values. Default is 0.5.}

\item{p}{weighted sum parameter for utility scores.Default is 0.5, i.e. equal phis.}

\item{beta}{F score type.}
}
\value{
a numerical value.
}
\description{
Calculation of weighted errors for weighted
regression models
}
\details{
Calculates weighted errors. They can be used as a performance or
model selection metric in imbalanced regression when phis are determined
same as train data. An example is given below. Implemented methods are

- WSSE: weighted sum of squared errors.

- WMSE: weighted mean squared error.

- WMAD: weighted mean absolute deviation.

- WMAPE: weighted mean absolute percentage error.

- MU: Mean utility.

- NMU: Normalized mean utility.

- precision for regression

- recall for regression

- Fbeta score for regression. Default beta is 1.
}
\examples{

x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

x_train = x[1:70]
y_train = y[1:70]

x_test = x[71:100]
y_test = y[71:100]

### resampling with SMOTER to predict rare values better
m_SMOTER <- SMOTER(x = x_train, y = y_train)

data_train <- data.frame(x = m_SMOTER$x_new, y = m_SMOTER$y_new)
m_lm <- lm(y~., data = data_train)

pred_test <- predict(m_lm, newdata = data.frame(x = x_test))

### phis for test data
m_rel <- relevance_PCHIP(y = y_train, y_new = y_test)

### weighted sum of squares
SSE_weighted <- WSSE(truth = y_test, pred = pred_test, phi = m_rel$rel)
SSE_weighted
### weighted mean squared error
MSE_weighted <- WMSE(truth = y_test, pred = pred_test, phi = m_rel$rel)
MSE_weighted

### weighted mean absolute deviation
MAD_weighted <- WMAD(truth = y_test, pred = pred_test, phi = m_rel$rel)
MAD_weighted

### weighted mean absolute percentage error
MAPE_weighted <- WMAPE(truth = y_test, pred = pred_test, phi = m_rel$rel)
MAPE_weighted

}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
