% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_density.R
\name{relevance_density}
\alias{relevance_density}
\title{Relevance function based on density}
\usage{
relevance_density(
  y,
  y_new = y,
  h = NULL,
  alpha = 1,
  bw_method = "ucv",
  epsilon = 1e-10
)
}
\arguments{
\item{y}{target values.}

\item{y_new}{new target values for calculating relevance values. Default is y.}

\item{h}{bandwith value for kernel density estimation. Default is NULL. If
NULL, it is estimated based on \code{method}.}

\item{alpha}{coefficient of density. Takes value in \eqn{(0,\infty)}. Higher
values mean less rare values. Lower values mean more rare values. Default is 1.}

\item{bw_method}{method to be used in estimating bandwidth. Default is unbiased
cross validation.}

\item{epsilon}{the lowest possible relevance value. Default is \eqn{1e-10}.}
}
\value{
a vector of relevance values.
}
\description{
Relevance value generation using kernel density estimation
}
\details{
\code{relevance_density} is a relevance function. It is used to determine
rare and not rare values in a regression problem. It calculates kernel densities
and inverse of densities are considered relevance values (Steininger et al., 2021).
}
\examples{

### regression dataset
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

### relevance values
rel <- relevance_density(y)

}
\references{
Steininger, M., Kobs, K., Davidson, P., Krause, A., & Hotho, A. (2021).
Density-based weighting for imbalanced regression. Machine Learning, 110,
2187-2211.
}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
