% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOTER.R
\name{SMOTER}
\alias{SMOTER}
\title{SMOTE for Regression}
\usage{
SMOTER(
  x,
  y,
  thresh_rel = 0.5,
  phi = NULL,
  perc_ov_lower = NULL,
  perc_ov_upper = NULL,
  perc_un = NULL,
  k = 5,
  method = "ucv"
)
}
\arguments{
\item{x}{feature matrix or dataframe. Only numeric variables for now.}

\item{y}{target variable.}

\item{thresh_rel}{threshold to determine rare values. Default is 0.5.}

\item{phi}{Relevance values for each sample. Default is NULL. If NULL,
inverse of kernel density estimation is used.}

\item{perc_ov_lower}{percentage of oversampling of lower rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_ov_upper}{percentage of oversampling of upper rare samples. If
NULL, it is automatically determined by algorithm. Cannot be lower than 1.}

\item{perc_un}{percentage of undersampling of non-rare samples. If NULL, it
is automatically determined by algorithm. Cannot be higher than 1.}

\item{k}{number of neighbors for links}

\item{method}{bandwith estimation method used in relevance value calculation.
Default is "ucv".}
}
\value{
an list object which includes:
 \item{x_new}{SMOTEd feature matrix}
 \item{y_new}{SMOTEd target variable}
 \item{x_syn}{Synthetic feature matrix}
 \item{y_syn}{Synthetic target variable}
 \item{phi}{relevance function values for y}
 \item{y}{original y to be used to calculate new target values}
 \item{h}{bandwidth value used to calculate densities for relevance values}
}
\description{
\code{SMOTER} applies undersampling and oversampling to
imbalanced regression data sets.
}
\details{
Despite its name, it can undersample and oversample imbalanced data sets. It is
used to better estimate the rare values in regression. Algorithm is a little
different to UBL package and more similar to original paper (Torgo et al., 2013).
Relevance values are the inverse of kernel density estimations using unbiased
cross-validation bandwidth.There are three classes: lower rare,
not rare and upper rare. Lower rare class is samples that satisfy \eqn{\phi > treshold}
and \eqn{y < \tilde{y}}. Upper rare class is samples that satisfy \eqn{\phi > treshold}
and \eqn{y > \tilde{y}}. Other samples are considered as not rare.
\eqn{\tilde{y}} is median of \eqn{y}.
}
\examples{
x <- rnorm(100)
err <- rnorm(100)
y <- 2 + x^2 + err

m_SMOTER <- SMOTER(x = x, y = y)

plot(x, y)
plot(m_SMOTER$x_new, m_SMOTER$y_new)

}
\references{
Torgo, L., Ribeiro, R. P., Pfahringer, B., & Branco, P. (2013, September).
Smote for regression. In Portuguese conference on artificial intelligence
(pp. 378-389). Berlin, Heidelberg: Springer Berlin Heidelberg.
}
\author{
Fatih Sağlam, saglamf89@gmail.com
}
